Sparse Arrays
### [Challenge Name: Sparse Arrays](/challenges/sparse-arrays)


There is a collection of $N$ strings ( There can be multiple occurences of a particular string ). Each string's length is no more than $20$ characters. There are also $Q$ queries. For each query, you are given a string, and you need to find out how many times this string occurs in the given collection of $N$ strings.