Print the Elements of a Linked List
### [Challenge Name: Print the Elements of a Linked List](/challenges/print-the-elements-of-a-linked-list)


This challenge is part of a [MyCodeSchool](http://www.youtube.com/mycodeschool) tutorial track and is accompanied by a [video lesson](http://www.youtube.com/embed/vcQIFT79_50?theme=light).

----

If you're new to *linked lists*, this is a great exercise for learning about them. Given a pointer to the *head* node of a linked list, print its elements in order, one element per line. If the head pointer is null (indicating the list is empty), don’t print anything.