Delete a Node
### [Challenge Name: Delete a Node](/challenges/delete-a-node-from-a-linked-list)


<sub>This challenge is part of a tutorial track by [MyCodeSchool](http://www.youtube.com/mycodeschool) and is accompanied by a video lesson.</sub>


You’re given the pointer to the head node of a linked list and the position of a node to delete. Delete the node at the given position and return the head node. A position of 0 indicates head, a position of 1 indicates one node away from the head and so on. The list may become empty after you delete the node.

**Input Format**  
You have to complete the <code>Node* Delete(Node* head, int position)</code> method which takes two arguments - the head of the linked list and the position of the node to delete. You should NOT read any input from stdin/console. <code>position</code> will always be at least 0 and less than the number of the elements in the list.

**Output Format**  
Delete the node at the given position and <code>return</code> the head of the updated linked list. Do NOT print anything to stdout/console.

**Sample Input**

1 --> 2 --> 3 --> NULL,  position =  0  
1 --> NULL ,  position = 0
    
**Sample Output**

    2 --> 3 --> NULL
    NULL

**Explanation**  
1. 0th position is removed, 1 is deleted from the list.  
2. Again 0th position is deleted and we are left with empty list. 

**Video lesson**

<iframe width="560" height="315" src="//www.youtube.com/embed/Y0n86K43GO4?theme=light" frameborder="0" allowfullscreen></iframe>