Find Merge Point of Two Lists
### [Challenge Name: Find Merge Point of Two Lists](/challenges/find-the-merge-point-of-two-joined-linked-lists)


This challenge is part of a tutorial track by [MyCodeSchool](http://www.youtube.com/mycodeschool)

----

Given pointers to the head nodes of $2$ linked lists that merge together at some point, find the Node where the two lists merge. It is guaranteed that the two head Nodes will be different, and neither will be NULL.

In the diagram below, the two lists converge at Node `x`:

	[List #1] a--->b--->c
                         \
                          x--->y--->z--->NULL
                         /
         [List #2] p--->q
     
Complete the `int FindMergeNode(Node* headA, Node* headB)` method so that it finds and returns the data value of the Node where the two lists merge.